messaging between clients is fully functional
Spent more time making it look pretty... client is in the terminal lol.
updated socket on to take parsed message and if it is chat to broadcast it instead of simply echo it back.
client side is an ugly amalgamation that is super simple post and get set up with a little bit of stuff to make it look pretty. is using readLine for IO then relies on net object called client that is linked to port 7777 to post and get from server and other clients