modified the server a little, changed ports, UE client changes
the changes to the ue client have made it much easier to run and is much more easy to follow. the names of the vars need to be changed for even better understanding. but is now more like the simple, get bytes and append, if it doesnt find a new line re roll and append until a new line. if there are some bunk characters before after the JSOn remove them before parsing